--- js/avatars.js
@@ -1,21 +1,0 @@
-import { getPosts } from './data.js';
-
-const pictureTemplate = document.querySelector('#picture').content.querySelector('.picture');
-const container = document.querySelector('.pictures');
-
-const fragment = document.createDocumentFragment();
-
-getPosts.forEach((photo) => {
-  const thumbnail = pictureTemplate.cloneNode(true);
-  const image = thumbnail.querySelector('.picture__img');
-
-  image.src = photo.url;
-  image.alt = photo.description;
-
-  thumbnail.querySelector('.picture__comments').textContent = photo.comments.length;
-  thumbnail.querySelector('.picture__likes').textContent = photo.likes;
-
-  fragment.appendChild(thumbnail);
-});
-
-container.appendChild(fragment);
