+++ js/posts.js
@@ -1,0 +1,23 @@
+const pictureTemplate = document.querySelector('#picture').content.querySelector('.picture');
+const container = document.querySelector('.pictures');
+const fragment = document.createDocumentFragment();
+
+const render = (dataList) => {
+  dataList.forEach((photo) => {
+    const thumbnail = pictureTemplate.cloneNode(true);
+    const image = thumbnail.querySelector('.picture__img');
+
+    image.src = photo.url;
+    image.alt = photo.description;
+
+    thumbnail.querySelector('.picture__comments').textContent = photo.comments.length;
+    thumbnail.querySelector('.picture__likes').textContent = photo.likes;
+
+    fragment.appendChild(thumbnail);
+  });
+
+  container.appendChild(fragment);
+};
+
+
+export { render };
